requirements.txt
crewai
litellm
python-dotenv
pydantic
fastapi>=0.110.0
uvicorn>=0.29.0